+++ src/SleepHunter.App/Settings/UserSettings.cs
@@ -25,7 +25,8 @@ namespace SleepHunter.Settings
         private bool saveMacroStates;
         private string selectedTheme;
 
-        private PlayerSortOrder clientSortOrder = PlayerSortOrder.LoginTime;
+        private PlayerSortOrder clientSortOrder =
+            PlayerSortOrder.LaunchOrder;
         private bool suppressLoginNotification = true;
         private bool applyModifiersKeyFix = true;
         private bool allowAltToShowGroundItems = true;
@@ -584,7 +585,7 @@ namespace SleepHunter.Settings
             SaveMacroStates = true;
 
             SelectedTheme = ColorThemeManager.Instance.DefaultTheme?.Name ?? "Default";
-            clientSortOrder = PlayerSortOrder.LoginTime;
+            clientSortOrder = PlayerSortOrder.LaunchOrder;
             InventoryIconSize = 46;
             SkillIconSize = 46;
             InventoryGridWidth = 12;